# Speed up Python module import