--- a/interfaces/cython/setup.cfg.in
+++ b/interfaces/cython/setup.cfg.in
@@ -42,6 +42,7 @@ include_package_data = True
 install_requires =
     numpy >= 1.12.0
     ruamel.yaml >= 0.15.34
+    packaging
 python_requires = >=@py_min_ver_str@
 packages =
     cantera