--- a/site_scons/buildutils.py
+++ b/site_scons/buildutils.py
@@ -11,7 +11,10 @@ import time
 import shutil
 import enum
 from pathlib import Path
-from pkg_resources import parse_version
+try:
+    from packaging.version import parse as parse_version
+except ImportError:
+    from pkg_resources import parse_version
 import logging
 from typing import TYPE_CHECKING
 from collections.abc import Mapping as MappingABC
@@ -1352,7 +1355,10 @@ def get_pip_install_location(
     root = quoted(root) if root is not None else None
     install_script = textwrap.dedent(f"""
         from pip import __version__ as pip_version
-        from pkg_resources import parse_version
+        try:
+            from packaging.version import parse as parse_version
+        except ImportError:
+            from pkg_resources import parse_version
         import pip
         import json
         pip_version = parse_version(pip_version)