<!-- Thanks for contributing code! Please include a description of your change and check your pull request against the list below. For further questions, refer to the contributing guide (https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md). -->

**Changes proposed in this pull request**

<!-- Provide a clear and concise description of changes and/or features introduced in this pull request. -->

- Replace use of `pkg_resources` with faster alternatives
- Defer loading `pandas` and `scipy` until they are used

**If applicable, provide an example illustrating new features this pull request is introducing**

This reduces the time it takes to import the Cantera Python module by ~70%. This can be examined by running
```
PYTHONPATH=build/python python3 -X importtime -c 'import cantera'
```

On my system (Mac M1 Pro), where the total import time is around 261 ms, the biggest contributors to this are:
- `pandas`: 104 ms
- `cantera._utils`: 80 ms, which encompasses loading `scipy.sparse`
- `pkg_resources`: 55 ms
- `numpy`: 42 ms

With this PR, the total import time is reduced to 80 ms. Of the items mentioned above, `numpy` remains the same, `cantera._utils` decreases to 8 ms, and the others are eliminated.

<!-- A minimal, complete, and reproducible example demonstrating features introduced by this pull request. See https://stackoverflow.com/help/minimal-reproducible-example for additional suggestions on how to create such an example. -->

**Checklist**

- [x] The pull request includes a clear description of this code change
- [x] Commit messages have short titles and reference relevant issues
- [x] Build passes (`scons build` & `scons test`) and unit tests address code coverage
- [x] Style & formatting of contributed code follows [contributing guidelines](https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md)
- [x] The pull request is ready for review